0.03

+++ popisinventara/users/routes.py
@@ -30,12 +30,22 @@ def login():
 
 @users.route("/user_list", methods=['GET', 'POST'])
 def user_list():
+    if not current_user.is_authenticated:
+        flash('Da biste pristupili ovoj stranici treba da budete ulogovani.', 'danger')
+        return redirect(url_for('users.login'))
+    if current_user.authorization != 'admin':
+        flash('Nemate dozvolu za prikaz korisnika.', 'danger')
+        return redirect(url_for('main.home'))
     users = User.query.all()
-    return render_template('user_list.html', users=users)
+    number_of_admins = User.query.filter_by(authorization='admin').count()
+    return render_template('user_list.html', users=users, number_of_admins=number_of_admins )
 
 
 @users.route("/register_user", methods=['GET', 'POST'])
 def register_user():
+    if not current_user.is_authenticated:
+        flash('Da biste pristupili ovoj stranici treba da budete ulogovani.', 'danger')
+        return redirect(url_for('users.login'))
     if current_user.authorization != 'admin':
         flash('Nemate dozvolu za registraciju korisnika.', 'danger')
         return redirect(url_for('main.home'))
@@ -43,7 +53,13 @@ def register_user():
     surname = request.form.get('surname').capitalize()
     authorization = request.form.get('authorization')
     email = request.form.get('email')
-    new_user = User(name=name, surname=surname, authorization=authorization, email=email, school_id=1, password='test')
+    hashed_password = bcrypt.generate_password_hash(request.form.get('password')).decode('utf-8')
+    new_user = User(name=name, 
+                    surname=surname, 
+                    authorization=authorization, 
+                    email=email, 
+                    school_id=1, 
+                    password=hashed_password)
     db.session.add(new_user)
     db.session.commit()
     flash(f'Korisnik {name} {surname} je uspešno registrovan!', 'success')
@@ -52,6 +68,9 @@ def register_user():
 
 @users.route("/edit_user", methods=['GET', 'POST'])
 def edit_user():
+    if not current_user.is_authenticated:
+        flash('Da biste pristupili ovoj stranici treba da budete ulogovani.', 'danger')
+        return redirect(url_for('users.login'))
     if current_user.authorization != 'admin':
         flash('Nemate dozvolu za izmene korisnika.', 'danger')
         return redirect(url_for('main.home'))
@@ -60,6 +79,7 @@ def edit_user():
     user.name = request.form.get('edit_name').capitalize()
     user.surname = request.form.get('edit_surname').capitalize()
     user.email = request.form.get('edit_email')
+    user.authorization = request.form.get('edit_authorization')
     db.session.commit()
     
     flash(f'Profil korisnika {user.name} {user.surname} je uspešno izmenjen', 'success')
@@ -68,10 +88,13 @@ def edit_user():
 
 @users.route("/delete_user", methods=['GET', 'POST'])
 def delete_user():
+    if not current_user.is_authenticated:
+        flash('Da biste pristupili ovoj stranici treba da budete ulogovani.', 'danger')
+        return redirect(url_for('users.login'))
     if current_user.authorization != 'admin':
         flash('Nemate dozvolu za brisanje korisnika.', 'danger')
         return redirect(url_for('main.home'))
-    user_id = request.form.get('user_id')
+    user_id = request.form.get('delete_user_id')
     user = User.query.filter_by(id=user_id).first()
     db.session.delete(user)
     db.session.commit()
